Move SnapLike import location
