+++ plonk/visualize/plots.py
@@ -6,7 +6,7 @@ import matplotlib as mpl
 import numpy as np
 from numpy import ndarray
 
-from ..snap.snap import SnapLike
+from ..snap import SnapLike
 from .interpolation import Extent
